Create a graphical testing application and fix AABB-AABB collisions
I tried to use geometry softwares to visually check whether vinage is right
when it detects a collision, but they aren't made for this specific task so
it is hard to use them.
This application will be used to test Vinage, and it already made it
possible to fix an AABB-AABB bug!