Fix SAT-based collision function
This is used to test ObbObb collisions. ObbObb collisions have not been tested
yet in a wrapped universe (because the graphical tester has no support) and are
thus unlikely to work in that case.

Also, remove a bug when geometric objects are used as prototypes. When the
object's properties were set, it was actually its prototype's properties which
were set.